requirements.txt
flask==3.0.0
anthropic==0.18.1
pillow==10.2.0
openpyxl==3.1.2
python-dotenv==1.0.0
werkzeug==3.0.1
